Fixed the pom.xml to eliminate redundant dependencies.
When running in Eclipse, the various javax.servlet jars were
colliding with each other causing problems. The conflicts were a result
of the javax.servlet jars from hadoop, hbase and dropwizard. Removing the
ones from Hadoop and HBase and keeping the ones in DropWizard seems to fix
the problem. Just something to be aware of if we bring in more hadoop related
dependencies.

Reviewed by: shashir