REST-10: Ensure that the /rows resource streams properly in all cases.
Clean up some of the internal APIs so that when a single row is requested by
specifying the eid parameter, it is returned to the client in a fashion
consistent with if multiple rows were returned. Previously, this single row was
not terminated with the \r\n making clients (i.e. ruby gem) fail.

Reviewer: lsheng