REST-12. Emit JSON that is properly navigable instead of a list of cells.
The first iteration of REST emitted rows where each row contained a list of
cells. This made it hard for clients to navigate properly having to linearly
search the list of cells to find one that matched a particular family/qualifer.
Now the output consists of a map of maps allowing direct access to the list of
versioned cells via the family and qualifier keys (e.g.
output["family"]["qualifier][0]["value"]).

Cells are represented as a timestamp/value tuple where the value itself is a
properly navigable JSON object.

reviewer: lsheng