REST-45. Add the writer schema to the output cells.
Fixes a problem with being able to POST what you GET by adding the writer schema
(id) to each cell being written. This way, you get a representation of the cell
as it was written and then you are guaranteed that on write, you can write back
properly using the writer schema for the cell. Previously the writer schema was
the schema declared in the layout which is not going to be valid in layout-1.3.

reviewer: lsheng, shashir