REST-47. Moved the ruby gem to the server codebase.
As we are going to expand the number of clients that we provide, it makes more
sense to move them into a subfolder of the umbrella KijiREST codebase. This will
help ensure that clients are updated and that we don't have a ton of little
repositories for each possible KijiREST client.

Reviewer: lsheng