REST-55. Fix the POM and assembly files to not produce a fat jar.
The problem with a fat jar is that dependencies used are masked by the far jar
creation and any collisions between dependencies at runtime could cause
problems. While these should be eliminated from the pom, it's much easier to
debug a collision by removing jars from the lib/ folder than recompiling a jar
file. This change will remove the creation of a far jar instead opting for
copying the runtime dependencies to the lib/ folder. The startup scripts behave
the same as it relies on lib/* for classpath entries which previously included.

Also preparing for 1.0 release by up-ing the POM version.

Reviewer: lsheng, shashir